Safari-609.1.11.1/LayoutTests/fast/xpath/nsresolver-bad-object-expected.txt

CONSOLE MESSAGE: line 19: XPathNSResolver does not have a lookupNamespaceURI method.
SUCCESS
